Move swap from <ranges> to <concepts>
