+++ include/concepts/include/vccc/concepts.hpp
@@ -34,6 +34,7 @@
 #include "vccc/concepts/signed_integral.hpp"
 #include "vccc/concepts/strict_weak_order.hpp"
 #include "vccc/concepts/subtractable.hpp"
+#include "vccc/concepts/swap.hpp"
 #include "vccc/concepts/swappable.hpp"
 #include "vccc/concepts/totally_ordered.hpp"
 #include "vccc/concepts/unsigned_integral.hpp"
