+++ include/ranges/include/vccc/ranges.hpp
@@ -21,7 +21,6 @@
 #include "vccc/ranges/range.hpp"
 #include "vccc/ranges/size.hpp"
 #include "vccc/ranges/sized_range.hpp"
-#include "vccc/ranges/swap.hpp"
 #include "vccc/ranges/swap_ranges.hpp"
 #include "vccc/ranges/view.hpp"
 #include "vccc/ranges/view_interface.hpp"
